% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{Cell clustering (dispatcher)

TODO: Add hierarchical clustering}
\usage{
clustering(sce, method = "graph", ...)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object}

\item{method}{String, method of clustering}

\item{...}{Additional parameters for specified clustering method}
}
\value{
\code{\link{SingleCellExperiment}} object
}
\description{
Cell clustering (dispatcher)

TODO: Add hierarchical clustering
}
