% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countQC.R
\name{countQC}
\alias{countQC}
\title{Main function for read count quality control (dispatcher).

TODO: Look into inheritParams tag}
\usage{
countQC(sce, method, rm.genes = TRUE, ...)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object}

\item{method}{Character specifying which quality control method to use}

\item{rm.genes}{Remove mitochondrial and ribosomal genes}

\item{...}{Additional parameters for method function}
}
\value{
\code{\link{SingleCellExperiment}} object
}
\description{
Main function for read count quality control (dispatcher).

TODO: Look into inheritParams tag
}
