% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReduced.R
\name{dimReduced}
\alias{dimReduced}
\title{Dimensionality reduction (dispatcher)}
\usage{
dimReduced(sce, seed = SEED, method = "vanilla", subset.name = "subset", ...)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object}

\item{seed}{Integer, random number generator initial state}

\item{method}{String, method of dimensionality reduction}

\item{subset.name}{String, name of subset to be used}

\item{...}{Additional parameters for dispatched functions}
}
\value{
\code{\link{SingleCellExperiment}} object
}
\description{
Dimensionality reduction (dispatcher)
}
