% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countQC.R
\name{.adaptiveFilter}
\alias{.adaptiveFilter}
\title{Filters cells based on quartiles.}
\usage{
.adaptiveFilter(sce)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object}
}
\value{
\code{\link{SingleCellExperiment}} object
}
\description{
Generate quartiles for mitochondrial percent, library size, and number of
genes detected. Cells in the first quartile of any feature is removed.
}
