% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countQC.R
\name{.between}
\alias{.between}
\title{Determine if a value is within a given range.}
\usage{
.between(criteria.data, .range)
}
\arguments{
\item{criteria.data}{A numeric vector}

\item{.range}{A numeric vector of length 2.}
}
\value{
A logical vector of cells to be removed.
}
\description{
Determine if a value is within a given range.
}
