% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countQC.R
\name{.build.criteria.patterns}
\alias{.build.criteria.patterns}
\title{Build criteria name patterns corresponding to the percent column.}
\usage{
.build.criteria.patterns(criteria.names)
}
\arguments{
\item{criteria.names}{String vector of criteria names.}
}
\value{
List of modified criteria names. The list is appended with \code{BASE.CRITERIA},
which include general criteria (ex. \code{sum}).
}
\description{
Build a string pattern for criteria that corresponds to the calculated
percentage that criteria occupy within a cell (ie. mitochondrial count proportion).
}
