% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countQC.R
\name{.build.quartile}
\alias{.build.quartile}
\title{Builds the quartile ranges for a given criteria.}
\usage{
.build.quartile(criteria.column)
}
\arguments{
\item{criteria.column}{Numeric vector for one QC criteria.}
}
\value{
Named vector corresponding the quartile ranges.
}
\description{
Builds the quartile ranges for a given criteria.
}
