% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countQC.R
\name{.build.subsets}
\alias{.build.subsets}
\title{Builds logical vector defining the a subset (ie. criteria).}
\usage{
.build.subsets(sce, gene.symbol.patterns)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object.}

\item{gene.symbol.patterns}{A list of gene symbol patterns.}
}
\value{
A list of gene symbol patterns.
}
\description{
Builds a logical vector that corresponds to matches to the gene symbol pattern
(ex. '^MT-') specified by the user. These logical vectors will define which genes (rows)
will be used to compute the criteria values (ie. mitochondrial proportion).
}
