% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normCounts.R
\name{.deconvolution}
\alias{.deconvolution}
\title{Pooled-counts based normalization.}
\usage{
.deconvolution(
  sce,
  seed = SEED,
  method = "igraph",
  min.size = 10,
  use.ranks = FALSE,
  bsparam = BiocSingular::IrlbaParam(),
  min.cells = 150,
  min.proportion = 0.3,
  min.window.size = 101,
  min.mean = 0.1,
  steps = 5,
  initial.pool = 21
)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object}

\item{seed}{Random number generator initial state}

\item{method}{String, clustering method}

\item{min.size}{Integer, minimum number of clusters}

\item{use.ranks}{Logical, use rank matrix for clustering}

\item{bsparam}{Algorithm to be used for PCA}

\item{min.cells}{Integer, TODO}

\item{min.proportion}{Numeric, TODO}

\item{min.window.size}{Integer, TODO}

\item{min.mean}{Numeric, TODO}

\item{steps}{Integer, TODO}

\item{initial.pool}{Integer, TODO}
}
\value{
\code{\link{SingleCellExperiment}} object
}
\description{
Counts from multiple cells are pooled to increase the size of the counts for
accurate size factor estimation. Pooling is done by clustering of cells, using the
\code{igraph} method. Each cluster is normalized separately and size factors are rescaled
to make clusters comparable. Clustering is performed on a PCA subspace, and must #' be seeded.
}
