% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReduced.R
\name{.explained.variance}
\alias{.explained.variance}
\title{Percent of explained variance}
\usage{
.explained.variance(sce)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object}
}
\value{
Numeric vector
}
\description{
Percent of explained variance
}
