% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReduced.R
\name{.findElbowPoint}
\alias{.findElbowPoint}
\title{Find the elbow point in the curve of variance explained by each successive PC. This can be used to determine the number of PCs to retain.}
\usage{
.findElbowPoint(variance)
}
\arguments{
\item{variance}{Numeric vector containing the variance explained by each PC.
Should be monotonic decreasing.}
}
\value{
An integer scalar specifying the number of PCs at the elbow point.
}
\description{
Find the elbow point in the curve of variance explained by each successive PC. This can be used to determine the number of PCs to retain.
}
\details{
Find the elbow point in the curve of variance explained by each successive PC. This can be used to determine the number of PCs to retain.
}
\author{
Aaron Lun
}
