% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffExp.R
\name{.findMarkers}
\alias{.findMarkers}
\title{Find markers within a dataset}
\usage{
.findMarkers(sce)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object}
}
\description{
Find markers within a dataset
}
