% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countQC.R
\name{.fixedFilter}
\alias{.fixedFilter}
\title{Filters cells based on fixed thresholds.}
\usage{
.fixedFilter(sce, nmads = 3, mito.range = c(1, 60), ribo.range = c(0, 60))
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object}

\item{nmads}{Median absolute deviation}

\item{mito.range}{Min/max vector for mitochondrial percentage}

\item{ribo.range}{Min/max vector for ribosomal percentage}
}
\value{
SingleCellExperiment object
}
\description{
Filters cells based on fixed thresholds.
}
\details{
Removes cells (columns) that do not meet the paramaterized criteria. These cells
can be referred to as low-quality libraries that may arise due to cell damage, or
failure during library preparation (ex. drop-out events)

The median absolute deviation is a metric for variability, As the name suggests,
it is the deviation of the absolute deviation (sum of differences between point X
and the median) of a vector of values. This method should not be influenced by
extreme outlier values, and can be used instead of z-scores. Assumes a normal
distribution, without consideration of outlier events.

Library size filtering performed by \code{\link[scater]{isOutlier}}, looking at the 'lower'
end of the distribution, in the log scale. Number of detected genes is also
filtered under the same criteria.
}
