% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{.generate.layout}
\alias{.generate.layout}
\title{Generates layout for force-directed visualization.

TODO: Bad practice, call function from string...}
\usage{
.generate.layout(sce, graph, layout)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object}

\item{graph}{Graph object (\code{\link[igraph]{make_graph}})}

\item{layout}{String, layout method}
}
\value{
\code{\link{SingleCellExperiment}} object
}
\description{
Generates layout for force-directed visualization.

TODO: Bad practice, call function from string...
}
