% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{.get.cluster.method}
\alias{.get.cluster.method}
\title{Get cluster method from igraph}
\usage{
.get.cluster.method(cluster)
}
\arguments{
\item{cluster}{String, community detection method}
}
\value{
closure
}
\description{
Get cluster method from igraph
}
