% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countQC.R
\name{.get.criteria.data}
\alias{.get.criteria.data}
\title{Get the columns for the criteria that will be used for quality control.}
\usage{
.get.criteria.data(sce, criteria.names)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object.}

\item{criteria.names}{A string vector with the user-defined criteria (ex. mito).}
}
\value{
S4Vectors object, each column representing one criteria.
}
\description{
Get the columns for the criteria that will be used for quality control.
}
