% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countQC.R
\name{.get.criteria.ranges}
\alias{.get.criteria.ranges}
\title{Get criteria ranges.}
\usage{
.get.criteria.ranges(criteria)
}
\arguments{
\item{criteria}{A list of lists.}
}
\value{
List of numeric vectors.
}
\description{
Get the second elements of the \code{criteria} list of lists, which corresponds to
a criteria range. The range is the lower and upper bounds a criteria value in a cell.
}
