% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countQC.R
\name{.get.gene.symbol.patterns}
\alias{.get.gene.symbol.patterns}
\title{Get criteria patterns.}
\usage{
.get.gene.symbol.patterns(criteria)
}
\arguments{
\item{criteria}{A list of lists.}
}
\value{
List of gene symbol patterns.
}
\description{
Get the first elements of the \code{criteria} list of lists, which corresponds to
a gene symbol pattern (ex. '^MT-').
}
