% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{.graph}
\alias{.graph}
\title{Graph-based clustering.

TODO: Separate graph generation and community detection.
TODO: Assess cluster separation}
\usage{
.graph(
  sce,
  k,
  type = "number",
  cluster = "walktrap",
  layout = "fr",
  dimred = "PCA"
)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object}

\item{k}{Integer, number of nearest neighbours (resolution)}

\item{type}{String, edge weighting scheme}

\item{cluster}{String, community detection method}

\item{layout}{String, method for force-directed graph layout}

\item{dimred}{String, dimensionality reduced subspace}
}
\value{
\code{\link{SingleCellExperiment}} object
}
\description{
Generate a k-nearest neighbours graph for downstream community detection.
\code{\link[igraph]{cluster_walktrap}} detects densely connected subgraphs via a
random walk. The intuition for this algorithm is that random walks tend to get
trapped in regions of high density.
}
