% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReduced.R
\name{.hvg.subset}
\alias{.hvg.subset}
\title{Subset highly variably genes}
\usage{
.hvg.subset(sce, subset.name)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object}

\item{subset.name}{String, name of subset}
}
\value{
Logical vector
}
\description{
Subset highly variably genes
}
