% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countQC.R
\name{.is.base.criteria}
\alias{.is.base.criteria}
\title{Determine if a criteria name is from the base criteria.}
\usage{
.is.base.criteria(criteria.name)
}
\arguments{
\item{criteria.name}{A string for the name of a criteria}
}
\value{
logical
}
\description{
Determine if a criteria name is from the base criteria.
}
