% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{.kmeans}
\alias{.kmeans}
\title{k-means clustering

TODO: Assess cluster gap statistic
TODO: Assess cluster separation}
\usage{
.kmeans(sce, k, field)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object}

\item{k}{Integer, number of clusters (centroids)}

\item{field}{String, dimensionality-reduced subspace to use}
}
\value{
\code{\link{SingleCellExperiment}} object
}
\description{
k-means clustering

TODO: Assess cluster gap statistic
TODO: Assess cluster separation
}
\details{
Clustering cells based on centroids, minimizing for within-cluster sum of squares.
In other words, it groups cells that are closests to a given centroid with could be
any given cell or a "proto"-cell.

This method is biased towards normally distributed, spherical, clusters.
}
