% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countQC.R
\name{.legacy.flag.cells}
\alias{.legacy.flag.cells}
\title{Flag cells to be removes from \code{\link{SingleCellExperiment}} object.}
\usage{
.legacy.flag.cells(criteria.column, criteria.name, .ranges, nmads)
}
\arguments{
\item{criteria.column}{Numeric vector for one QC criteria.}

\item{criteria.name}{Name of the criteria currnetly being operated on.}

\item{nmads}{The median absolute deviation for \code{\link[scuttle]{isOutlier}}.}
}
\value{
Logical matrix of cell to be removed.
}
\description{
Flag cells to be removes from \code{\link{SingleCellExperiment}} object.
}
\details{
Individual QC criteria columns are being processed in this function. For example,
the \code{criteria.column} vector would correspond to the number of detected genes column.
}
