% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countQC.R
\name{.legacy}
\alias{.legacy}
\title{The go to quality control protocol in our lab.}
\usage{
.legacy(criteria.data, ...)
}
\arguments{
\item{criteria.data}{An S4Vector with quality control criteria of interest.}

\item{...}{Additional parameters for \code{.legacy.flag.cells}.}
}
\value{
Logical vector, flagging cells to be removed.
}
\description{
The go to quality control protocol in our lab.
}
\details{
The legacy protocol is a mixture of fixed and adaptive thresholds,
specific to a given criteria.

Mitochondrial and ribosomal gene percentages are subject to user-defined boundaries,
which are specified within the list of lists discussed in \code{countQC}. Cells whose
percentages are within a given range are kept.

Library size (\code{sum}) and number of detected genes (\code{detected}) are
flagged by the \code{\link[scuttle]isOutlier} function. The method relies on median
absolute deviation, which is the median distance any given cell has to the median
value. This method should not be influenced by extreme outliers. The outlier detection
is parameterized to look at the lower tail (ex. small library sizes) in a log2 scale.
}
