% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReduced.R
\name{.load.variance.model}
\alias{.load.variance.model}
\title{Load variance model generated during feature selection.}
\usage{
.load.variance.model()
}
\description{
Load variance model generated during feature selection.
}
