% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normCounts.R
\name{.max.window.size}
\alias{.max.window.size}
\title{Maximum window size.}
\usage{
.max.window.size(ncolumns, min.cells, min.window.size, proportion)
}
\arguments{
\item{ncolumns}{Integer, number of columns}

\item{min.cells}{Integer, minimum number of cells}

\item{min.window.size}{Numeric, minimum upper limit of window size}

\item{proportion}{Numeric, proportion of cells to be pooled given ncolumns}
}
\value{
Integer
}
\description{
The number of cells to be pooled together.
}
