% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normCounts.R
\name{.normVanilla}
\alias{.normVanilla}
\title{Scaling normalization by library size.}
\usage{
.normVanilla(sce, ...)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object}

\item{...}{\code{\link[scater]{librarySizeFactors}} arguments}
}
\value{
\code{\link{SingleCellExperiment}} object
}
\description{
Scaling normalization by library size.
}
\details{
May not yield accurate normalized expression values for downstream analysis, but
sufficient for exploratory analysis.

Each cells is a library, and the size of the library is the total sum of counts
across all genes in that cell. The library size factor is directly proportional to
library size, relative to the mean size factor across all cells equaling 1.

Normalized values will be on the same scale as the original counts

Any upregulation is met with an eqaul magnitude downregulation, since this method
assumes that the dataset is perfectly normal. This prevents composition effects.
}
