% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countQC.R
\name{.outside.range}
\alias{.outside.range}
\title{Determines if numeric is outside a given range, exclusively.}
\usage{
.outside.range(numerics, range, operator)
}
\arguments{
\item{numerics}{Numeric vector}

\item{range}{Numeric vector representing the min/max range}

\item{operator}{Comparison (logical) operator function}
}
\value{
Logical vector
}
\description{
Determines if numeric is outside a given range, exclusively.
}
\details{
Given a numeric vector, determine which values are outside a range
specified by the 'range' vector. The min/max comparison are combined using a
logical operator (ex. \code{|} or \code{&}).
}
