% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReduced.R
\name{.pca}
\alias{.pca}
\title{Perform principal component analysis on cells}
\usage{
.pca(sce, hvariable.genes, ncomponents, infer, variance.model)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object}

\item{hvariable.genes}{String, location of HVGs}

\item{ncomponents}{Integer, number of components to retain}

\item{infer}{Logical, if the number of components should be inferred}

\item{variance.model}{DataFrame, Generated variance model from feature selection}
}
\value{
\code{\link{SingleCellExperiment}} object
}
\description{
Perform principal component analysis on cells
}
