% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelect.R
\name{.poisson}
\alias{.poisson}
\title{Model gene variance by fitting into poisson distribution.}
\usage{
.poisson(sce, seed = SEED, ...)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object}

\item{seed}{Random number generator initial state}

\item{...}{Additional \code{\link[scran]{modelGeneVarByPoisson}} arguments}
}
\value{
\code{\link{SingleCellExperiment}}
}
\description{
Model gene variance by fitting into poisson distribution.
}
\details{
Without a standard (ie. spike-in) the technical noise can be modeled by our
understanding of the data-generation process. We assume that technical noise
manifests as a Poisson distribution.

This model assumes that most genes do not vary due to biological heterogeneity.
}
