% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countQC.R
\name{.quartile.flag.cells}
\alias{.quartile.flag.cells}
\title{Determine if cells are in the first quartile range.}
\usage{
.quartile.flag.cells(criteria.column)
}
\arguments{
\item{criteria.column}{Numeric vector for one QC criteria.}
}
\value{
Logical matrix, flagging cells to be removed.
}
\description{
Determine if cells are in the first quartile range.
}
