% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countQC.R
\name{.quartile}
\alias{.quartile}
\title{Generate quartiles for given numeric vector.}
\usage{
.quartile(data)
}
\arguments{
\item{data}{Numeric vector}
}
\value{
Named vector of quartile ranges
}
\description{
Generate quartiles for given numeric vector.
}
