% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countQC.R
\name{.quickPerCellQCMetrics}
\alias{.quickPerCellQCMetrics}
\title{Identify low QC cells with frequent QC metrics.}
\usage{
.quickPerCellQCMetrics(sce, ...)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}}}

\item{...}{TODO}
}
\value{
DataFrame
}
\description{
Identify low QC cells with frequent QC metrics.
}
