% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countQC.R
\name{.reduce.flag.matrix}
\alias{.reduce.flag.matrix}
\title{Collapses a logical matrix, row-wise, into a logical vector.}
\usage{
.reduce.flag.matrix(mtx)
}
\arguments{
\item{mtx}{A logical matrix.}
}
\value{
A logical vector.
}
\description{
Performs a \code{\link{|}} operation using row-wise sums. If the value is greater than
\code{0}, then the result is \code{TRUE}. All \code{TRUE} elements correspond to cells that are to be
discarded.
}
\seealso{
\link{Reduce} for a conceptual understanding.
}
