% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReduced.R
\name{.reducedVanilla}
\alias{.reducedVanilla}
\title{Conventional dimensionality reduction (PCA, tSNE, and UMAP)}
\usage{
.reducedVanilla(
  sce,
  subset.name,
  pca.ncomponents = 30,
  pca.infer = NULL,
  ...,
  tsne.perplexity = 20,
  tsne.ncomponents = 3,
  umap.min.dist = 0.5,
  umap.nneighbours = 15,
  umap.metric = "euclidean",
  umap.ncomponents = 3
)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object}

\item{subset.name}{String, name of subset to be used}

\item{pca.ncomponents}{Integer, number of components to retain}

\item{pca.infer}{String, method of inference for ncomponents}

\item{...}{Additional arguments for .pca}

\item{tsne.perplexity}{Numeric, perplexity (size of search radius)}

\item{tsne.ncomponents}{Integer, number of components}

\item{umap.min.dist}{Numeric, minimum distance between two nodes}

\item{umap.nneighbours}{Integer, number of neighbours}

\item{umap.metric}{String, distance metric}

\item{umap.ncomponents}{Integer, number of components}
}
\value{
\code{\link{SingleCellExperiment}} object
}
\description{
Performs PCA on read count matrix and selects user-defined, or
inferred, top number of components. The selected PCA subspace is then used
to initialize tSNE and UMAP algorithms.
}
