% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{.subcluster}
\alias{.subcluster}
\title{Subsclustering

TODO: Implement}
\usage{
.subcluster(sce, cluster)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object}

\item{cluster}{Integer, cluster of interest}
}
\description{
Subsclustering

TODO: Implement
}
