% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelect.R
\name{.vanilla}
\alias{.vanilla}
\title{Model gene variance by log2-transformed normalized counts.}
\usage{
.vanilla(sce, density.weight = TRUE, ...)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object}

\item{density.weight}{Logical, use of inverse density weights}

\item{...}{Additional \code{\link[scran]{modelGeneVar}} arguments}
}
\value{
\code{\link{SingleCellExperiment}}
}
\description{
Model gene variance by log2-transformed normalized counts.
}
\details{
Gene-base trend decomposition of technical and biological
components via fitting of mean-variance trend to model the variance of log2
expression profiles of each genes. This model assumes that most genes do not
vary due to biological heterogeneity.

Default arguments for \code{\link[scran]{modelGeneVar}} may lead to overfitting of the data,
when the high-abundance genes are also highly variable.
}
