% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.where}
\alias{.where}
\title{Find the index of strings that match a given pattern.}
\usage{
.where(pattern, strings, matcher)
}
\arguments{
\item{pattern}{A string pattern.}

\item{strings}{A vector of strings.}

\item{matcher}{A matching function (ex. \code{\link{grepl}})}
}
\value{
Vector of numerics corresponding to indexes where the pattern matched in the string vector.
}
\description{
Find the index of strings that match a given pattern.
}
