% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelect.R
\name{featureSelect}
\alias{featureSelect}
\title{Select highly variable genes based on per-gene variation model (dispatcher).}
\usage{
featureSelect(
  sce,
  n = NULL,
  prop = NULL,
  fdr.threshold = NULL,
  fit.trend = "poisson",
  subset.name = "subset"
)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object}

\item{n}{Integer, number of genes to select}

\item{prop}{Numeric, proportion of genes to select}

\item{fdr.threshold}{Numeric, false discovery rate threshold}

\item{fit.trend}{String, the trend tha data will be fit into}

\item{subset.name}{String, name highly variable gene subset}
}
\value{
\code{\link{SingleCellExperiment}} object
}
\description{
Select highly variable genes based on per-gene variation model (dispatcher).
}
\details{
Noise to signal ratio is a function of the number of highly variable genes
retained, given that the larger \code{n} is, the more biologically disinteresting
genes there will be in the final dataset.

Set \code{n} as the number of genes that you expect to be differentially expressed
in your system. Ballpark a number between 500 to 5,000 to start, if you do not
know how many genes you would expect to be differentially expressed.

Otherwise, select a false discovery rate (FDR) threshold (\code{fdr.threshold}),
where the proportion of false positive genes are expected to have an FDR
above the specified threshold. Number of genes will vary depending on the
false positive error rate and severity of multiple test correction. Genes
could also be ranked by their p-value, and obtain the top hits.

Selecting for genes above a certain variance will maximize noise, but minimize
bias. This is default setting in \code{\link[scran]{getTopHVGs}}.
}
