% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normCounts.R
\name{normCounts}
\alias{normCounts}
\title{Normalization of read counts (dispatcher)}
\usage{
normCounts(sce, method = "deconvolution", ...)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object}

\item{method}{String, normalization method}

\item{...}{Additional arguments specific to normalization method}
}
\value{
\code{\link{SingleCellExperiment}}
}
\description{
Normalization of read counts (dispatcher)
}
