% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving.R
\name{save.diagnostic}
\alias{save.diagnostic}
\title{TODO: Add documentation}
\usage{
save.diagnostic(obj, name)
}
\arguments{
\item{obj}{\code{\link{SingleCellExperiment}} object}

\item{name}{String, name of file}
}
\description{
TODO: Add documentation
}
