% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save.processed}
\alias{save.processed}
\title{TODO: Add documentation}
\usage{
save.processed(obj, name)
}
\arguments{
\item{obj}{\code{\link{SingleCellExperiment}} object}

\item{name}{String, name of file}
}
\description{
TODO: Add documentation
}
