% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countQC.R
\name{countQC}
\alias{countQC}
\title{Dispatcher function for quality control protocols.}
\usage{
countQC(sce, criteria, protocol = "legacy", rm.genes = TRUE, ...)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object.}

\item{criteria}{A list of a list containig details for each user-defined criteria.

The first element of a sublist must contain the gene symbol pattern to be searched.
The second element is only required if \code{legacy} protocol is being used, and it must contain
a vector of two numeric elements specifying the lower and upper bounds for the criteria value.

Common criteria include:
\itemize{
\item Mitochondrial gene percentage (ex. list(mito = list('^MT-', c(1, 60))))
\item Ribosomal gene percentage (ex. list(ribo = list('^RP[SL][\link{:digit:}], c(0, 60))))
}}

\item{protocol}{String specifying QC protocol.
Current protocols include:
\itemize{
\item 'legacy'
\item 'quartile'
\item 'scater'
}}

\item{rm.genes}{Determine if rows corresponding to QC criteria should be removed.}
}
\value{
\code{\link{SingleCellExperiment}} object
}
\description{
Perfoms user-specified quality control (QC) protocols on
\code{\link{SingleCellExperiment}} objects.
}
\details{
Implementated quality control methods are:
\itemize{
\item Fixed tresholds: User-defined threshold for specified QC metrics.
\item Quartile ranges: Removal of cells that fit within the first quartile range.
\item \code{\link[scater]{quikckPerCellQCMetrics}}: Relies on QC protocol implemented within \code{\link{scater}} .
}
}
