% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normCounts.R
\name{.deconvolution}
\alias{.deconvolution}
\title{Deconvolution-based read count normalization based on pooling of cells}
\usage{
.deconvolution(
  sce,
  clustering = list(method = "igraph", min.size = 10, use.ranks = FALSE, BSPARAM =
    BiocSingular::IrlbaParam()),
  deconv = list(assay = "counts", min.window.size = 21, max.window.size = list(ncells =
    150, size = 101, proportion = 0.3), steps = 5, min.mean = 0.1)
)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object.}

\item{clustering}{A named list of parameters for \code{\link[scran]{quickCluster}}.
\itemize{
\item \code{method}: The method of clustering to be performed (ie. \code{'igraph'}, \code{'hclust'})
\item \code{min.size}: The minimum number of clusters.
\item \code{use.ranks}: Whether or not to use the rank matrix for clustering.
\item \code{BSPARAM}: The PCA algorithm to be used, selected from \code{link{BiocSingular}}.
}}

\item{deconv}{A named list of parameters for \code{\link[scran]{computeSumFactors}}.
\itemize{
\item \code{assay}: The counts matrix to be used for normalization.
\item \code{min.window.size}: The minimum number of cells in one pool.
\item \code{max.window.size}: A list of parameters determining the maximum number of cells in a pool.
\itemize{
\item \code{ncells}: Number of cells in pool.
\item \code{size}: User-specified maximum window size.
\item \code{proportion}: Determining max size by proportion of cells
}
\item \code{steps}: Sliding window.
\item \code{min.mean}: The minimum mean per pool of cells.
}}
}
\value{
\code{\link{SingleCellExperiment}} object.
}
\description{
TODO
}
