% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countQC.R
\name{.quartile}
\alias{.quartile}
\title{Flag cells that are in the first quartile of given criteria.}
\usage{
.quartile(criteria.data)
}
\arguments{
\item{criteria.data}{An S4Vector with quality control criteria of interest.}
}
\value{
Logical vector, flagging cells to be removed.
}
\description{
Flag cells that are in the first quartile of given criteria.
}
\details{
This method does not rely on user-defined boundaries for criteria values.
Therefore, the user does \emph{not} have to specify a range.

The quartile protocol is adaptive. It will infer low quality cells from the data itself.
This could be beneficial since it removes human bias. However, the adaptive protocols
may remove more cells that necessary/desired. It is best to cross-reference this protocol
with the \code{legacy} implementation.
}
