% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countQC.R
\name{.scater}
\alias{.scater}
\title{Quick quality control using \code{\link[scater]quickPerCellQC}.}
\usage{
.scater(sce, ...)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment}} object.}

\item{...}{Additional parameters for \code{\link[scater]{quickPerCellQC}}.}
}
\description{
Quick quality control using \code{\link[scater]quickPerCellQC}.
}
