% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normCounts.R
\name{normCounts}
\alias{normCounts}
\title{Dispatcher function for normalization of read counts.}
\usage{
normCounts(sce, protocol, seed = 100, ...)
}
\description{
Dispatcher function for normalization of read counts.
}
