% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep.data}
\alias{prep.data}
\title{Prepare data for processing}
\usage{
prep.data(path10x, experiment.name)
}
\arguments{
\item{path10x}{Path to CellRanger output directory}

\item{experiment.name}{String, experiment identifier}
}
\value{
\code{\link{SingleCellExperiment}} object
}
\description{
Prepare data for processing
}
